requirements.txt
llama_cpp_python==0.3.13
numpy==2.2.6
pandas==2.3.1
python-dotenv==1.1.1
safetensors==0.5.3
tabulate==0.9.0
together==1.5.21
torch==2.7.1
transformers==4.53.2
typer==0.15.4
cffi~=1.17.1
typing_extensions~=4.12.2
requests~=2.32.4
openai~=1.77.0
tqdm~=4.67.1
yaml~=0.2.5
pyyaml~=6.0.2
pydantic~=2.11.7
annotated-types~=0.6.0
ipython~=9.1.0

